+++ sdk/resourcemanager/containerservice/armcontainerservice/options.go
@@ -22,6 +22,13 @@ type AgentPoolsClientBeginCreateOrUpdateOptions struct {
 	ResumeToken string
 }
 
+// AgentPoolsClientBeginDeleteMachinesOptions contains the optional parameters for the AgentPoolsClient.BeginDeleteMachines
+// method.
+type AgentPoolsClientBeginDeleteMachinesOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
 // AgentPoolsClientBeginDeleteOptions contains the optional parameters for the AgentPoolsClient.BeginDelete method.
 type AgentPoolsClientBeginDeleteOptions struct {
 	// ignore-pod-disruption-budget=true to delete those pods on a node without considering Pod Disruption Budget